requirements.txt
tweepy==4.14.0
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
SQLAlchemy==2.0.27
redis==5.0.1
celery==5.3.6
Pillow==10.2.0
pytesseract==0.3.10
python-binance==1.0.19
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
aiohttp==3.9.3 